requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
Flask-Migrate==4.0.5

